Fix test failing due to __getattr__ in module
The deprecated_handler() function is called by module level __getattr__
functions. It raised an ImportError when a module wasn't an env.
This breaks getattr's behaviour when it is called with a default
value. It expects __getattr__ to raise an AttributeError if the attr
is not found.

Pytest was apparently updated to call getattr(obj, key, default)
for several values of key. The import error isn't caught by getattr,
breaking the test.

Changing ImportError to AttributeError fixes this. The missing
attribute raises the correct error, which is caught by getattr,
which sets the value to the default and everything works as
intended.