Fix bug in SB3 tutorial ActionMask
SB3ActionMaskWrapper.step() is intended to be compatible with Gymnansium's
interface where step() returns observation, reward, termination, truncation, info

This was implemented using the last() function. But this returns the values for
the current agent, not the agent that just acted as Gymnasium would.

Among other things, this trains on the opponent's reward, encouraging bad play.

The function now returns the reward, termination, truncation, info values for the
agent that just acted. It still returns the observation for the next agent since
it is used to determine the next action.