Fix pyright errors in AECOrderEnforcingIterator
The problem stemmed from the iterator being based on AECEnv
but accessing env._has_updated which only exists in the
OrderEnforcingWrapper.

Pyright's error is correct based on the types given in the code.
However, the usage of the class is correct. This is a workaround
that uses a more precise type and convinces pyright that the code
is correct.