Ignore pygame deprecation warnings during testing
pygame is using a deprecated packaging api. Since it has been raising warnings for years, it seems unlikely that the pygame devs are interested in fixing that any time soon. This change ignores those warnings during testing. If pygame devs ever see the light and correct this, these filters should be removed.