Fix incorrect termination in cooperative_pong
The ball was not checked for exiting both vertical and horizontal
dimensions at the same time. This led to case where the ball was out of
bounds but the game continued for an extra step.

This change corrects that and increments the game version since the
behaviour is different.

Fixes #1286